Nova_Search/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.1
langchain-openai==0.0.5
openai==1.3.5
faiss-cpu==1.12.0
elasticsearch==8.10.0
pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
clip-interrogator==0.6.0
torch==2.1.0
torchvision==0.16.0
pytest==7.4.3
requests==2.31.0
